overhauling infer script
